+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Version 2.0
 
+- `Data Export` add capabilities to highlight SQL keywords, add comments in SOQL (removed on query run), and pretty format the SOQL (button) [666](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/666) [359](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/359) (contribution by [Nicolas Greard](https://github.com/ngreardSF))
 - Fix popup/detail record name field [issue #274](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/274) (contribution by [Nicolas Greard](https://github.com/ngreardSF))
 - `Data Export` Fix Column order not respected when performing subqueries [#598](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/598)
 - `Accessibility` Improve screen reader support: fix AlertBanner silent announcements, convert inspector button and tooltip trigger to semantic elements, add popup iframe title issues [#1107](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/1107) & [#1108](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/1108) (contribution by [akj](https://github.com/akj))